fix(downloader): preserve retained write retry state
