+++ backend/tests/unit/protocol/downloader.test.ts
@@ -1,183 +1,10 @@
 import { describe, it, expect, beforeEach, afterEach } from 'bun:test';
 import { Downloader } from '../../../src/protocol/downloader.ts';
-import { ChunkDownloader, type ChunkDownloaderDeps, type RetryInfo } from '../../../src/protocol/chunk-downloader.ts';
-import { PeerManager } from '../../../src/protocol/peer-manager.ts';
-import { PauseController } from '../../../src/protocol/pause-controller.ts';
-import { ProgressReporter } from '../../../src/protocol/progress-reporter.ts';
 import type { IStoredLISH, LISHid, ChunkID } from '@shared';
 import { CodedError, ErrorCodes } from '@shared';
 import type { MissingChunk } from '../../../src/lish/data-server.ts';
-import type { ChunkSlot } from '../../../src/db/lishs-chunks.ts';
-import type { FileVerificationProgress } from '../../../src/db/lishs-verification.ts';
 import { MockNetwork } from '../helpers/mock-network.ts';
-
-// ---------------------------------------------------------------------------
-// Mock LISHClient
-// ---------------------------------------------------------------------------
-
-type ChunkResult = Uint8Array | null | Error;
-type ManifestResult = IStoredLISH | null;
-
-interface ChunkVerifyResult {
-	valid: boolean;
-	actualHash: string;
-}
-
-class MockLISHClient {
-	requestChunkResult: ChunkResult = new Uint8Array(1024).fill(0xff);
-	requestManifestResult: ManifestResult = null;
-	requestManifestError: Error | null = null;
-	requestManifestCalls = 0;
-	closeCalled = false;
-	haveChunks: 'all' | ChunkID[] = 'all';
-	requestChunkCalls = 0;
-	chunkData: Map<ChunkID, Uint8Array> | null = null; // if set, serve per-chunkID (multi-chunk tests)
-
-	async requestChunk(_lishID: LISHid, chunkID: ChunkID): Promise<Uint8Array | null> {
-		this.requestChunkCalls++;
-		if (this.chunkData) return this.chunkData.get(chunkID) ?? null;
-		if (this.requestChunkResult instanceof Error) throw this.requestChunkResult;
-		return this.requestChunkResult;
-	}
-
-	async requestManifest(_lishID: LISHid): Promise<IStoredLISH | null> {
-		this.requestManifestCalls++;
-		if (this.requestManifestError) throw this.requestManifestError;
-		return this.requestManifestResult;
-	}
-
-	async close(): Promise<void> {
-		this.closeCalled = true;
-	}
-
-	/** Immediate teardown — records the same flag; tests assert the client was released. */
-	abort(): void {
-		this.closeCalled = true;
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Minimal DataServer mock
-// ---------------------------------------------------------------------------
-
-class MockDataServer {
-	missingChunks: MissingChunk[] = [];
-	allChunkCount = 0;
-	downloadedChunks = new Set<ChunkID>();
-	writtenChunks: Array<{ fileIndex: number; chunkIndex: number; data: Uint8Array }> = [];
-	addedLishs: IStoredLISH[] = [];
-	storedLishs = new Map<string, IStoredLISH>();
-	completeLishs = new Set<string>();
-
-	getMissingChunks(_lishID: LISHid): MissingChunk[] {
-		return [...this.missingChunks];
-	}
-
-	getAllChunkCount(_lishID: LISHid): number {
-		return this.allChunkCount;
-	}
-
-	// One slot per missing chunk — enough for run()'s duplicate-checksum grouping in tests.
-	getAllChunkSlots(_lishID: LISHid): ChunkSlot[] {
-		return this.missingChunks.map(c => ({ fileIndex: c.fileIndex, chunkIndex: c.chunkIndex, checksum: c.chunkID }));
-	}
-
-	getFileVerificationProgress(_lishID: LISHid): FileVerificationProgress[] {
-		return [];
-	}
-
-	isChunkDownloaded(_lishID: LISHid, chunkID: ChunkID): boolean {
-		return this.downloadedChunks.has(chunkID);
-	}
-
-	markChunkDownloaded(_lishID: LISHid, chunkID: ChunkID): void {
-		this.downloadedChunks.add(chunkID);
-		// Remove from missingChunks to mimic real DB behaviour
-		this.missingChunks = this.missingChunks.filter(c => c.chunkID !== chunkID);
-	}
-
-	writeChunkError: Error | null = null;
-	writeChunkErrorCount = 0; // how many more times to throw before succeeding
-	writeChunkOutcomes: Array<Error | null> | null = null; // per-call script (shift); null entry = success
-	async writeChunk(_dir: string, _lish: IStoredLISH, fileIndex: number, chunkIndex: number, data: Uint8Array): Promise<void> {
-		if (this.writeChunkOutcomes) {
-			const outcome = this.writeChunkOutcomes.shift() ?? null;
-			if (outcome) throw outcome;
-			this.writtenChunks.push({ fileIndex, chunkIndex, data });
-			return;
-		}
-		if (this.writeChunkError && this.writeChunkErrorCount > 0) {
-			this.writeChunkErrorCount--;
-			throw this.writeChunkError;
-		}
-		this.writtenChunks.push({ fileIndex, chunkIndex, data });
-	}
-
-	incrementDownloadedBytes(_lishID: LISHid, _bytes: number): void {
-		/* no-op for tests */
-	}
-
-	resetFileChunks(_lishID: LISHid, fileIndex: number): number {
-		// Re-add all chunks for this fileIndex to missingChunks
-		const lish = [...this.storedLishs.values()][0];
-		if (!lish?.files?.[fileIndex]) return 0;
-		const file = lish.files[fileIndex]!;
-		let resetCount = 0;
-		for (let i = 0; i < file.checksums.length; i++) {
-			const chunkID = file.checksums[i]! as ChunkID;
-			if (this.downloadedChunks.has(chunkID)) {
-				this.downloadedChunks.delete(chunkID);
-				this.missingChunks.push({ chunkID, fileIndex, chunkIndex: i });
-				resetCount++;
-			}
-		}
-		return resetCount;
-	}
-
-	add(lish: IStoredLISH): void {
-		this.addedLishs.push(lish);
-		this.storedLishs.set(lish.id, lish);
-	}
-
-	get(lishID: LISHid): IStoredLISH | null {
-		return this.storedLishs.get(lishID) ?? null;
-	}
-
-	isCompleteLISH(lish: IStoredLISH): boolean {
-		return this.completeLishs.has(lish.id);
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function makeLISH(overrides: Partial<IStoredLISH> = {}): IStoredLISH {
-	return {
-		id: 'test-lish-id-0001' as LISHid,
-		name: 'Test LISH',
-		created: new Date().toISOString(),
-		chunkSize: 1024 * 1024,
-		checksumAlgo: 'sha256',
-		files: [{ path: 'file.bin', size: 1024, checksums: ['abc123'] }],
-		directory: '/tmp/test-download',
-		...overrides,
-	};
-}
-
-function makeMissingChunk(chunkID: ChunkID, fileIndex = 0, chunkIndex = 0): MissingChunk {
-	return { chunkID, fileIndex, chunkIndex };
-}
-
-/** Access private members for test assertions. */
-function priv(d: Downloader): Record<string, unknown> {
-	return d as unknown as Record<string, unknown>;
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
+import { MockDataServer, MockLISHClient, makeLISH, makeMissingChunk, priv } from './downloader-test-helpers.ts';
 describe('Downloader – static speed limit', () => {
 	afterEach(() => {
 		// Reset static limit between tests
@@ -710,977 +537,6 @@ describe('Downloader – initFromManifest state transitions', () => {
 
 		const missing = priv(downloader)['missingChunks'] as MissingChunk[];
 		expect(missing).toHaveLength(2);
-	});
-});
-
-// ---------------------------------------------------------------------------
-// Path traversal protection (safePath)
-// ---------------------------------------------------------------------------
-
-describe('Downloader – path traversal protection', () => {
-	let ds: MockDataServer;
-	let net: MockNetwork;
-
-	beforeEach(() => {
-		ds = new MockDataServer();
-		net = new MockNetwork();
-	});
-
-	function makeDownloader(downloadDir = '/tmp/safe-downloads/12345'): Downloader {
-		return new Downloader(downloadDir, net as never, ds as never, 'net-001');
-	}
-
-	function callSafePath(downloader: Downloader, relativePath: string): string {
-		const fa = (downloader as unknown as { fileAllocator: { safePath: (p: string) => string } }).fileAllocator;
-		return fa.safePath(relativePath);
-	}
-
-	// --- Legitimate paths (should pass) ---
-
-	it('allows simple filename', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'movie.mkv')).not.toThrow();
-	});
-
-	it('allows nested path', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'video/movie.mkv')).not.toThrow();
-	});
-
-	it('allows deeply nested path', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'a/b/c/d/e/file.txt')).not.toThrow();
-	});
-
-	it('allows path with spaces', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'my folder/my file.txt')).not.toThrow();
-	});
-
-	it('allows path with unicode characters', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'složka/soubor-čěšřž.txt')).not.toThrow();
-	});
-
-	// --- Basic traversal attacks (must block) ---
-
-	it('blocks ../ at start', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, '../evil.txt')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks ../../ double traversal', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, '../../evil.txt')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks ../../../ triple traversal', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, '../../../etc/passwd')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks ../ in middle of path', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'subdir/../../evil.txt')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks deeply nested traversal that escapes', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'a/b/c/../../../../evil.txt')).toThrow('Path traversal blocked');
-	});
-
-	// --- Encoded/obfuscated traversal attempts (must block) ---
-
-	it('blocks backslash traversal on Windows', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, '..\\evil.txt')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks mixed slash traversal', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, '..\\..\\evil.txt')).toThrow('Path traversal blocked');
-	});
-
-	// --- Boundary: traversal that stays inside (should pass) ---
-
-	it('allows subdir/../same-level (stays inside downloadDir)', () => {
-		const dl = makeDownloader();
-		// subdir/../file.txt resolves to downloadDir/file.txt — still inside
-		expect(() => callSafePath(dl, 'subdir/../file.txt')).not.toThrow();
-	});
-
-	it('allows a/b/../b/file.txt (stays inside downloadDir)', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, 'a/b/../b/file.txt')).not.toThrow();
-	});
-
-	// --- Targeted attack scenarios ---
-
-	it('blocks settings.json overwrite attack', () => {
-		const dl = makeDownloader('/app/data/downloads/12345');
-		expect(() => callSafePath(dl, '../../settings.json')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks .ssh/authorized_keys attack', () => {
-		const dl = makeDownloader('/home/user/libershare/downloads/12345');
-		expect(() => callSafePath(dl, '../../../../../.ssh/authorized_keys')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks /etc/passwd attack via deep traversal', () => {
-		const dl = makeDownloader('/app/data/downloads/12345');
-		expect(() => callSafePath(dl, '../../../../../../etc/passwd')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks database overwrite attack', () => {
-		const dl = makeDownloader('/app/.node1/downloads/12345');
-		expect(() => callSafePath(dl, '../../libershare.db')).toThrow('Path traversal blocked');
-	});
-
-	// --- Absolute path injection (must block) ---
-
-	it('blocks absolute path on Unix', () => {
-		const dl = makeDownloader('/tmp/safe');
-		expect(() => callSafePath(dl, '/etc/passwd')).toThrow('Path traversal blocked');
-	});
-
-	// --- Edge cases ---
-
-	it('blocks bare ..', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, '..')).toThrow('Path traversal blocked');
-	});
-
-	it('blocks empty-segment traversal /../', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, './../evil')).toThrow('Path traversal blocked');
-	});
-
-	it('allows . (current dir)', () => {
-		makeDownloader();
-		// resolve(downloadDir, '.') = downloadDir — but startsWith(downloadDir + sep) fails
-		// because it equals downloadDir exactly, without trailing sep
-		// This is an edge case — '.' as a file path is unusual but harmless
-		// safePath would throw here, but that's acceptable (no real file is named '.')
-	});
-
-	it('allows ./file.txt', () => {
-		const dl = makeDownloader();
-		expect(() => callSafePath(dl, './file.txt')).not.toThrow();
-	});
-});
-
-// ---------------------------------------------------------------------------
-// Chunk integrity verification
-// ---------------------------------------------------------------------------
-
-describe('Downloader – chunk integrity verification', () => {
-	// Utility: compute real sha256 for a given Uint8Array
-	function sha256(data: Uint8Array): string {
-		const hasher = new Bun.CryptoHasher('sha256');
-		hasher.update(data);
-		return hasher.digest('hex');
-	}
-
-	// Simulate the inline verification logic extracted from peerLoop
-	function verifyChunk(data: Uint8Array, expectedChunkID: string, algo: string): ChunkVerifyResult {
-		const hasher = new Bun.CryptoHasher(algo as any);
-		hasher.update(data);
-		const actualHash = hasher.digest('hex');
-		return { valid: actualHash === expectedChunkID, actualHash };
-	}
-
-	// --- Basic integrity scenarios ---
-
-	it('accepts chunk with correct sha256 hash', () => {
-		const data = new Uint8Array([1, 2, 3, 4, 5]);
-		const expectedHash = sha256(data);
-		const result = verifyChunk(data, expectedHash, 'sha256');
-		expect(result.valid).toBe(true);
-	});
-
-	it('rejects chunk with wrong hash — random data', () => {
-		const data = new Uint8Array([1, 2, 3, 4, 5]);
-		const result = verifyChunk(data, 'deadbeef'.repeat(8), 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	it('rejects chunk with truncated hash', () => {
-		const data = new Uint8Array([1, 2, 3, 4, 5]);
-		const correctHash = sha256(data);
-		const result = verifyChunk(data, correctHash.slice(0, 32), 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	it('rejects empty data when non-empty hash expected', () => {
-		const expected = sha256(new Uint8Array([1, 2, 3]));
-		const result = verifyChunk(new Uint8Array(0), expected, 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	it('accepts empty data with correct empty hash', () => {
-		const emptyData = new Uint8Array(0);
-		const expected = sha256(emptyData);
-		const result = verifyChunk(emptyData, expected, 'sha256');
-		expect(result.valid).toBe(true);
-	});
-
-	// --- Attack: bit-flip in chunk data ---
-
-	it('detects single bit flip in chunk data', () => {
-		const data = new Uint8Array(1024).fill(0xaa);
-		const expected = sha256(data);
-		// Flip one bit
-		const corrupted = new Uint8Array(data);
-		corrupted[512] = 0xab;
-		const result = verifyChunk(corrupted, expected, 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	// --- Attack: peer sends all zeros ---
-
-	it('detects all-zero replacement attack', () => {
-		const realData = new Uint8Array(1024 * 1024); // 1MB
-		for (let i = 0; i < realData.length; i++) realData[i] = i % 256;
-		const expected = sha256(realData);
-		// Attacker sends all zeros
-		const zeroData = new Uint8Array(1024 * 1024);
-		const result = verifyChunk(zeroData, expected, 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	// --- Attack: peer sends different valid chunk (swapped chunks) ---
-
-	it('detects chunk swap attack — peer sends chunk B instead of chunk A', () => {
-		const chunkA = new Uint8Array([10, 20, 30]);
-		const chunkB = new Uint8Array([40, 50, 60]);
-		const hashA = sha256(chunkA);
-		// Peer sends chunkB data but we expect hashA
-		const result = verifyChunk(chunkB, hashA, 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	// --- Attack: peer sends shorter data ---
-
-	it('detects truncated chunk data (peer sends partial chunk)', () => {
-		const fullData = new Uint8Array(1024).fill(0xff);
-		const expected = sha256(fullData);
-		// Peer sends only half
-		const partial = fullData.slice(0, 512);
-		const result = verifyChunk(partial, expected, 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	// --- Attack: peer sends longer data (padding attack) ---
-
-	it('detects padded chunk data (peer appends extra bytes)', () => {
-		const realData = new Uint8Array([1, 2, 3, 4]);
-		const expected = sha256(realData);
-		// Peer appends extra data
-		const padded = new Uint8Array(8);
-		padded.set(realData);
-		padded.set([5, 6, 7, 8], 4);
-		const result = verifyChunk(padded, expected, 'sha256');
-		expect(result.valid).toBe(false);
-	});
-
-	// --- Multiple algorithms ---
-
-	it('verifies with sha512 algorithm', () => {
-		const data = new Uint8Array([1, 2, 3]);
-		const hasher = new Bun.CryptoHasher('sha512');
-		hasher.update(data);
-		const expected = hasher.digest('hex');
-		const result = verifyChunk(data, expected, 'sha512');
-		expect(result.valid).toBe(true);
-	});
-
-	it('rejects sha256 hash when algo is sha512', () => {
-		const data = new Uint8Array([1, 2, 3]);
-		const wrongAlgoHash = sha256(data); // sha256 hash
-		const result = verifyChunk(data, wrongAlgoHash, 'sha512'); // but algo=sha512
-		expect(result.valid).toBe(false);
-	});
-
-	it('verifies with blake2b256 algorithm', () => {
-		const data = new Uint8Array([1, 2, 3]);
-		const hasher = new Bun.CryptoHasher('blake2b256' as any);
-		hasher.update(data);
-		const expected = hasher.digest('hex');
-		const result = verifyChunk(data, expected, 'blake2b256');
-		expect(result.valid).toBe(true);
-	});
-
-	// --- Corruption counter & peer banning logic ---
-
-	it('MAX_CORRUPT_CHUNKS is 3', () => {
-		// Constant lives on ChunkDownloader after orchestrator refactor
-		expect((ChunkDownloader as any).MAX_CORRUPT_CHUNKS).toBe(3);
-	});
-
-	it('simulates peer ban after 3 corrupt chunks', () => {
-		const corruptCount = new Map<string, number>();
-		const MAX = 3;
-		const peerID = '12D3KooWEvil';
-		const banned: string[] = [];
-
-		// Simulate 3 corrupt chunks from same peer
-		for (let i = 0; i < 3; i++) {
-			const count = (corruptCount.get(peerID) ?? 0) + 1;
-			corruptCount.set(peerID, count);
-			if (count >= MAX) banned.push(peerID);
-		}
-
-		expect(corruptCount.get(peerID)).toBe(3);
-		expect(banned).toContain(peerID);
-	});
-
-	it('does not ban peer with fewer than 3 corruptions', () => {
-		const corruptCount = new Map<string, number>();
-		const MAX = 3;
-		const peerID = '12D3KooWSemiHonest';
-
-		// Only 2 corrupt chunks
-		for (let i = 0; i < 2; i++) {
-			corruptCount.set(peerID, (corruptCount.get(peerID) ?? 0) + 1);
-		}
-
-		expect(corruptCount.get(peerID)).toBe(2);
-		expect(corruptCount.get(peerID)! < MAX).toBe(true);
-	});
-
-	it('tracks corruption independently per peer', () => {
-		const corruptCount = new Map<string, number>();
-		const MAX = 3;
-
-		// Peer A: 2 corruptions (no ban)
-		corruptCount.set('peerA', 2);
-		// Peer B: 3 corruptions (ban)
-		corruptCount.set('peerB', 3);
-		// Peer C: 0 corruptions (clean)
-
-		expect(corruptCount.get('peerA')! < MAX).toBe(true);
-		expect(corruptCount.get('peerB')! >= MAX).toBe(true);
-		expect(corruptCount.get('peerC') ?? 0).toBe(0);
-	});
-
-	// --- Requeue behavior ---
-
-	it('corrupt chunk is requeued for other peers', () => {
-		const queue = ['chunk1', 'chunk2', 'chunk3'];
-		let queueIdx = 0;
-
-		// Peer takes chunk1
-		const taken = queue[queueIdx++];
-		expect(taken).toBe('chunk1');
-
-		// Chunk1 fails verification — requeue
-		queue.push(taken!);
-
-		// Next peer takes chunk2 (not chunk1 again — chunk1 is at the end)
-		const next = queue[queueIdx++];
-		expect(next).toBe('chunk2');
-
-		// Eventually chunk1 comes back
-		const retry = queue[queueIdx + 1]; // skip chunk3
-		expect(retry).toBe('chunk1');
-	});
-
-	// --- Edge: correct hash but wrong data length (shouldn't happen with sha256 but test anyway) ---
-
-	it('two different data with same length produce different hashes', () => {
-		const a = new Uint8Array(100).fill(0x00);
-		const b = new Uint8Array(100).fill(0x01);
-		expect(sha256(a)).not.toBe(sha256(b));
-	});
-
-	// --- Large chunk (1MB) verification performance ---
-
-	it('verifies 1MB chunk in reasonable time', () => {
-		const data = new Uint8Array(1024 * 1024);
-		for (let i = 0; i < data.length; i++) data[i] = i % 256;
-		const expected = sha256(data);
-		const start = Date.now();
-		const result = verifyChunk(data, expected, 'sha256');
-		const elapsed = Date.now() - start;
-		expect(result.valid).toBe(true);
-		expect(elapsed).toBeLessThan(100); // sha256 of 1MB should be <100ms
-	});
-});
-
-// ---------------------------------------------------------------------------
-// All LISH-supported hash algorithms (per LISH_DATA_FORMAT.md)
-// ---------------------------------------------------------------------------
-
-describe('Downloader – chunk integrity with all LISH hash algorithms', () => {
-	const LISH_ALGOS = ['sha256', 'sha384', 'sha512', 'sha512-256', 'sha3-256', 'sha3-384', 'sha3-512', 'blake2b256', 'blake2b512', 'blake2s256'] as const;
-
-	function hashWith(algo: string, data: Uint8Array): string {
-		const h = new Bun.CryptoHasher(algo as any);
-		h.update(data);
-		return h.digest('hex');
-	}
-
-	function verifyChunk(data: Uint8Array, expectedChunkID: string, algo: string): ChunkVerifyResult {
-		const h = new Bun.CryptoHasher(algo as any);
-		h.update(data);
-		const actualHash = h.digest('hex');
-		return { valid: actualHash === expectedChunkID, actualHash };
-	}
-
-	const testData = new Uint8Array(4096);
-	for (let i = 0; i < testData.length; i++) testData[i] = (i * 7 + 13) % 256;
-
-	for (const algo of LISH_ALGOS) {
-		it(`${algo}: accepts correct hash`, () => {
-			const expected = hashWith(algo, testData);
-			const result = verifyChunk(testData, expected, algo);
-			expect(result.valid).toBe(true);
-		});
-
-		it(`${algo}: rejects wrong data`, () => {
-			const expected = hashWith(algo, testData);
-			const bad = new Uint8Array(testData);
-			bad[0] = bad[0]! ^ 0xff; // flip first byte
-			const result = verifyChunk(bad, expected, algo);
-			expect(result.valid).toBe(false);
-		});
-
-		it(`${algo}: rejects hash from different algo`, () => {
-			// Use sha256 hash but verify with this algo (unless it IS sha256)
-			const wrongHash = hashWith('sha256', testData);
-			if (algo === 'sha256') return; // skip — same algo
-			const result = verifyChunk(testData, wrongHash, algo);
-			// sha512-256 and sha256 produce same-length hashes but different values
-			expect(result.valid).toBe(false);
-		});
-
-		it(`${algo}: produces correct hash length`, () => {
-			const hash = hashWith(algo, testData);
-			const expectedLengths: Record<string, number> = {
-				sha256: 64,
-				sha384: 96,
-				sha512: 128,
-				'sha512-256': 64,
-				'sha3-256': 64,
-				'sha3-384': 96,
-				'sha3-512': 128,
-				blake2b256: 64,
-				blake2b512: 128,
-				blake2s256: 64,
-			};
-			expect(hash.length).toBe(expectedLengths[algo]!);
-		});
-	}
-
-	// Cross-algo collision test: same data, different algos → different hashes
-	it('all algos produce unique hashes for same data', () => {
-		const hashes = LISH_ALGOS.map(algo => hashWith(algo, testData));
-		const unique = new Set(hashes);
-		expect(unique.size).toBe(LISH_ALGOS.length);
-	});
-
-	// Verify 1MB chunk performance for all algos
-	it('all algos hash 1MB chunk under 200ms each', () => {
-		const bigData = new Uint8Array(1024 * 1024);
-		for (let i = 0; i < bigData.length; i++) bigData[i] = i % 256;
-		for (const algo of LISH_ALGOS) {
-			const start = Date.now();
-			hashWith(algo, bigData);
-			const elapsed = Date.now() - start;
-			expect(elapsed).toBeLessThan(200);
-		}
-	});
-});
-
-// ---------------------------------------------------------------------------
-// Inline retry tests
-// ---------------------------------------------------------------------------
-
-describe('Downloader — inline ENOENT recovery', () => {
-	let dataServer: MockDataServer;
-	let network: MockNetwork;
-	let downloader: Downloader;
-
-	beforeEach(() => {
-		dataServer = new MockDataServer();
-		network = new MockNetwork();
-		downloader = new Downloader('/tmp/test', network as any, dataServer as any, ['net1']);
-	});
-
-	afterEach(() => {
-		downloader.destroy().catch(() => {});
-	});
-
-	it('retryCallback is set and callable', () => {
-		const calls: any[] = [];
-		downloader.setRetryCallback(info => calls.push(info));
-		expect(calls.length).toBe(0);
-	});
-
-	it('has fileReallocAttempts and writeRetryCount fields', async () => {
-		// chunkDownloader is created in initFromManifest; retry state lives there now
-		await downloader.initFromManifest(makeLISH());
-		const cd = priv(downloader)['chunkDownloader'] as { fileReallocAttempts: Map<number, number>; writeRetryCount: number };
-		expect(cd.fileReallocAttempts).toBeInstanceOf(Map);
-		expect(cd.writeRetryCount).toBe(0);
-	});
-
-	it('MAX_FILE_REALLOC is 3', () => {
-		expect((ChunkDownloader as any).MAX_FILE_REALLOC).toBe(3);
-	});
-
-	it('MAX_WRITE_RETRIES is 5', () => {
-		expect((ChunkDownloader as any).MAX_WRITE_RETRIES).toBe(5);
-	});
-
-	it('enable() resets fileReallocAttempts and writeRetryCount', async () => {
-		const lish = makeLISH();
-		dataServer.add(lish);
-		dataServer.allChunkCount = 1;
-		dataServer.missingChunks = [makeMissingChunk('abc123' as ChunkID)];
-		await downloader.initFromManifest(lish);
-		const cd = priv(downloader)['chunkDownloader'] as { fileReallocAttempts: Map<number, number>; writeRetryCount: number };
-		cd.fileReallocAttempts.set(0, 2);
-		cd.writeRetryCount = 3;
-		await downloader.enable();
-		expect(cd.fileReallocAttempts.size).toBe(0);
-		expect(cd.writeRetryCount).toBe(0);
-	});
-
-	it('writeChunkError in mock triggers error for testing', async () => {
-		dataServer.writeChunkError = Object.assign(new Error('ENOENT'), { code: 'ENOENT' });
-		dataServer.writeChunkErrorCount = 1;
-		await expect(dataServer.writeChunk('/tmp', makeLISH(), 0, 0, new Uint8Array(10))).rejects.toThrow('ENOENT');
-		// Second call succeeds
-		await expect(dataServer.writeChunk('/tmp', makeLISH(), 0, 0, new Uint8Array(10))).resolves.toBeUndefined();
-	});
-
-	it('resetFileChunks mock correctly resets downloaded chunks for a file', () => {
-		const lish = makeLISH({
-			files: [
-				{ path: 'a.bin', size: 1024, checksums: ['chunk-a1' as string, 'chunk-a2' as string] },
-				{ path: 'b.bin', size: 1024, checksums: ['chunk-b1' as string] },
-			],
-		});
-		dataServer.add(lish);
-		dataServer.downloadedChunks.add('chunk-a1' as ChunkID);
-		dataServer.downloadedChunks.add('chunk-a2' as ChunkID);
-		dataServer.downloadedChunks.add('chunk-b1' as ChunkID);
-		const resetCount = dataServer.resetFileChunks(lish.id, 0);
-		expect(resetCount).toBe(2);
-		expect(dataServer.downloadedChunks.has('chunk-a1' as ChunkID)).toBe(false);
-		expect(dataServer.downloadedChunks.has('chunk-a2' as ChunkID)).toBe(false);
-		expect(dataServer.downloadedChunks.has('chunk-b1' as ChunkID)).toBe(true);
-	});
-
-	it('resetFileChunks returns 0 for file with no downloaded chunks', () => {
-		const lish = makeLISH();
-		dataServer.add(lish);
-		const resetCount = dataServer.resetFileChunks(lish.id, 0);
-		expect(resetCount).toBe(0);
-	});
-});
-
-describe('Downloader — destroy drain', () => {
-	it('waits for active work and cancels download before its completion waiter is installed', async () => {
-		const dataServer = new MockDataServer();
-		const lish = makeLISH();
-		dataServer.completeLishs.add(lish.id);
-		const downloader = new Downloader('/tmp/test', new MockNetwork() as never, dataServer as never, ['net1']);
-		await downloader.initFromManifest(lish);
-
-		let workStarted!: () => void;
-		let releaseWork!: () => void;
-		const workEntered = new Promise<void>(resolve => {
-			workStarted = resolve;
-		});
-		const workBlocked = new Promise<void>(resolve => {
-			releaseWork = resolve;
-		});
-		(priv(downloader) as any).doWorkInternal = async () => {
-			workStarted();
-			await workBlocked;
-		};
-
-		const download = downloader.download();
-		const outcome = download.then(
-			() => null,
-			error => error
-		);
-		await workEntered;
-		let destroyed = false;
-		const destroying = downloader.destroy().then(() => {
-			destroyed = true;
-		});
-		await Promise.resolve();
-		expect(destroyed).toBe(false);
-
-		releaseWork();
-		await destroying;
-		const error = await outcome;
-
-		expect(error).toBeInstanceOf(CodedError);
-		expect((error as CodedError).code).toBe(ErrorCodes.DOWNLOAD_CANCELLED);
-		expect(priv(downloader)['downloadReject']).toBeUndefined();
-	}, 5000);
-
-	it('drains peer discovery and never publishes the next WANT after destroy', async () => {
-		const dataServer = new MockDataServer();
-		const lish = makeLISH();
-		dataServer.completeLishs.add(lish.id);
-		const network = new MockNetwork();
-		const topics: string[] = [];
-		let firstBroadcastStarted!: () => void;
-		let releaseFirstBroadcast!: () => void;
-		const firstBroadcastEntered = new Promise<void>(resolve => {
-			firstBroadcastStarted = resolve;
-		});
-		const firstBroadcastBlocked = new Promise<void>(resolve => {
-			releaseFirstBroadcast = resolve;
-		});
-		network.broadcast = async topic => {
-			topics.push(topic);
-			if (topics.length === 1) {
-				firstBroadcastStarted();
-				await firstBroadcastBlocked;
-				throw new Error('old network stopped');
-			}
-		};
-		const downloader = new Downloader('/tmp/test', network as never, dataServer as never, ['net-a', 'net-b']);
-		await downloader.initFromManifest(lish);
-
-		const discovery = (priv(downloader)['callForPeers'] as () => Promise<void>).call(downloader);
-		await firstBroadcastEntered;
-		let destroyed = false;
-		const destroying = downloader.destroy().then(() => {
-			destroyed = true;
-		});
-		await Promise.resolve();
-		expect(destroyed).toBe(false);
-
-		releaseFirstBroadcast();
-		await Promise.all([discovery, destroying]);
-
-		expect(topics).toEqual(['lish/net-a']);
-	}, 5000);
-});
-
-describe('Downloader — inline ENOSPC retry', () => {
-	let dataServer: MockDataServer;
-	let network: MockNetwork;
-	let downloader: Downloader;
-
-	beforeEach(() => {
-		dataServer = new MockDataServer();
-		network = new MockNetwork();
-		downloader = new Downloader('/tmp/test', network as any, dataServer as any, ['net1']);
-	});
-
-	afterEach(() => {
-		downloader.destroy().catch(() => {});
-	});
-
-	it('WRITE_RETRY_DELAY is 60000ms', () => {
-		expect((ChunkDownloader as any).WRITE_RETRY_DELAY).toBe(60000);
-	});
-
-	it('writePaused starts as false', () => {
-		const pc = priv(downloader)['pauseController'] as { writePaused: boolean };
-		expect(pc.writePaused).toBe(false);
-	});
-
-	it('waitIfWritePaused resolves immediately when not paused', async () => {
-		const pc = priv(downloader)['pauseController'] as { waitIfWritePaused: () => Promise<void> };
-		await expect(pc.waitIfWritePaused()).resolves.toBeUndefined();
-	});
-
-	it('resumeWrites resolves all pending waiters', async () => {
-		const pc = priv(downloader)['pauseController'] as {
-			pauseWrites: () => void;
-			resumeWrites: () => void;
-			waitIfWritePaused: () => Promise<void>;
-			writeResolvers: unknown[];
-			writePaused: boolean;
-		};
-		pc.pauseWrites();
-		const promises = [pc.waitIfWritePaused(), pc.waitIfWritePaused()];
-		// Yield a microtask so the awaited Promises register their resolvers before we assert.
-		await new Promise(r => setTimeout(r, 0));
-		expect(pc.writeResolvers.length).toBe(2);
-		pc.resumeWrites();
-		await Promise.all(promises);
-		expect(pc.writePaused).toBe(false);
-		expect(pc.writeResolvers.length).toBe(0);
-	});
-
-	it('stays paused until every holder releases', async () => {
-		const pc = priv(downloader)['pauseController'] as {
-			pauseWrites: () => void;
-			resumeWrites: () => void;
-			waitIfWritePaused: () => Promise<void>;
-			writeResolvers: unknown[];
-			writePaused: boolean;
-		};
-		// Two independent paths pause writes: the retained-write retry cycle and
-		// missing-file recovery. Whichever finishes first must not open the gate for
-		// the other, or a peer would claim a retry cycle someone else already owns.
-		pc.pauseWrites();
-		pc.pauseWrites();
-		const waiter = pc.waitIfWritePaused();
-		await new Promise(r => setTimeout(r, 0));
-		expect(pc.writeResolvers.length).toBe(1);
-
-		pc.resumeWrites();
-		await new Promise(r => setTimeout(r, 0));
-		expect(pc.writePaused).toBe(true); // one holder left
-		expect(pc.writeResolvers.length).toBe(1); // waiter still blocked
-
-		pc.resumeWrites();
-		await waiter;
-		expect(pc.writePaused).toBe(false);
-		expect(pc.writeResolvers.length).toBe(0);
-	});
-
-	it('ignores an unbalanced release instead of going negative', () => {
-		const pc = priv(downloader)['pauseController'] as { pauseWrites: () => void; resumeWrites: () => void; writePaused: boolean };
-		pc.resumeWrites(); // nobody was holding
-		expect(pc.writePaused).toBe(false);
-		pc.pauseWrites();
-		expect(pc.writePaused).toBe(true); // a stray release must not leave a negative debt
-		pc.resumeWrites();
-		expect(pc.writePaused).toBe(false);
-	});
-});
-
-// ---------------------------------------------------------------------------
-// Write-retry keeps the verified chunk in memory instead of re-downloading it
-// ---------------------------------------------------------------------------
-
-describe('ChunkDownloader — write-retry retains chunk in memory (no re-download)', () => {
-	function sha256Hex(data: Uint8Array): string {
-		const h = new Bun.CryptoHasher('sha256');
-		h.update(data);
-		return h.digest('hex');
-	}
-
-	let origDelay: number;
-
-	beforeEach(() => {
-		// Shrink the 60s retry pause so the loop runs in milliseconds under test.
-		origDelay = (ChunkDownloader as unknown as { WRITE_RETRY_DELAY: number }).WRITE_RETRY_DELAY;
-		(ChunkDownloader as unknown as { WRITE_RETRY_DELAY: number }).WRITE_RETRY_DELAY = 15;
-		Downloader.setMaxDownloadSpeed(0); // guard against a leftover throttle from other suites
-	});
-
-	afterEach(() => {
-		(ChunkDownloader as unknown as { WRITE_RETRY_DELAY: number }).WRITE_RETRY_DELAY = origDelay;
-	});
-
-	/**
-	 * Build a ChunkDownloader over real PeerManager/PauseController/ProgressReporter with a
-	 * single missing chunk. `failCount` write attempts throw ENOSPC before the write succeeds
-	 * (`Infinity` = never succeeds). The peer client counts how many times the chunk is fetched.
-	 */
-	function harness(failCount: number) {
-		const data = new Uint8Array(2048);
-		for (let i = 0; i < data.length; i++) data[i] = (i * 5 + 1) & 0xff;
-		const chunkID = sha256Hex(data) as ChunkID;
-
-		const ds = new MockDataServer();
-		const lish = makeLISH({ files: [{ path: 'file.bin', size: data.length, checksums: [chunkID] }] });
-		ds.add(lish);
-		ds.allChunkCount = 1;
-		ds.missingChunks = [makeMissingChunk(chunkID, 0, 0)];
-		ds.writeChunkError = Object.assign(new Error('ENOSPC'), { code: 'ENOSPC' });
-		ds.writeChunkErrorCount = failCount;
-
-		const client = new MockLISHClient();
-		client.requestChunkResult = data;
-
-		const peerManager = new PeerManager();
-		peerManager.setLishID(lish.id);
-		peerManager.tryAdd('peer-1', client as never, 'DIRECT');
-
-		const state = { disabled: false, destroyed: false };
-		const pauseController = new PauseController(
-			() => state.disabled,
-			() => state.destroyed
-		);
-		const progressReporter = new ProgressReporter();
-
-		const retries: RetryInfo[] = [];
-		const errors: Array<{ code: string; detail: string | undefined }> = [];
-
-		const deps: ChunkDownloaderDeps = {
-			lishID: lish.id,
-			downloadDir: '/tmp/test-dl',
-			abortSignal: new AbortController().signal,
-			dataServer: ds as never,
-			peerManager,
-			pauseController,
-			progressReporter,
-			fileAllocator: {} as never,
-			getLish: () => lish,
-			isDestroyed: () => state.destroyed,
-			isDisabled: () => state.disabled,
-			onSetError: (code, detail) => {
-				errors.push({ code, detail });
-				state.disabled = true; // mirror Downloader.setError so peer loops exit
-			},
-			onRetry: info => retries.push({ ...info }),
-			emitAllocProgress: () => {},
-		};
-
-		return { cd: new ChunkDownloader(deps), client, ds, retries, errors, chunkID, data, pauseController, deps };
-	}
-
-	it('write fails once → retries from the SAME buffer, no second network request', async () => {
-		const h = harness(1);
-		await h.cd.run();
-
-		expect(h.client.requestChunkCalls).toBe(1); // fetched from the network exactly once
-		expect(h.ds.writtenChunks).toHaveLength(1);
-		expect(h.ds.writtenChunks[0]!.data).toEqual(h.data); // identical bytes landed on disk
-		expect(h.ds.downloadedChunks.has(h.chunkID)).toBe(true);
-		expect(h.ds.missingChunks).toHaveLength(0); // download completed
-		expect(h.retries.some(r => !r.resolved)).toBe(true); // FE got a retry notification
-		expect(h.retries.some(r => r.resolved)).toBe(true); // …and a resolved one
-		expect(h.errors).toHaveLength(0);
-	});
-
-	it('onRetry throws while taking the recovery pause → hold is released, not leaked', async () => {
-		const h = harness(0);
-		// The write pause is a holder count, so a hold leaked by a throwing callback would
-		// never self-heal — every later peer loop would block on it for this Downloader's life.
-		h.ds.writeChunkOutcomes = [Object.assign(new Error('ENOENT'), { code: 'ENOENT' })];
-		h.deps.onRetry = () => {
-			throw new Error('callback blew up');
-		};
-		await h.cd.run();
-
-		expect((h.pauseController as unknown as { writePauseHolders: number })['writePauseHolders']).toBe(0);
-		expect(h.pauseController.writePaused).toBe(false);
-		expect(h.pauseController.progressPaused).toBe(false);
-	});
-
-	it('write fails several times → buffer held across all retries, still one fetch', async () => {
-		const h = harness(3);
-		await h.cd.run();
-
-		expect(h.client.requestChunkCalls).toBe(1);
-		expect(h.ds.writtenChunks).toHaveLength(1);
-		expect(h.ds.writtenChunks[0]!.data).toEqual(h.data);
-		expect(h.ds.downloadedChunks.has(h.chunkID)).toBe(true);
-	});
-
-	it('write never succeeds → onSetError after MAX_WRITE_RETRIES, no re-download', async () => {
-		const h = harness(Number.POSITIVE_INFINITY);
-		await h.cd.run();
-
-		expect(h.client.requestChunkCalls).toBe(1); // buffer dropped only after giving up — never re-fetched
-		expect(h.errors).toHaveLength(1);
-		expect(h.errors[0]!.code).toBe(ErrorCodes.DISK_FULL);
-		expect(h.ds.downloadedChunks.has(h.chunkID)).toBe(false);
-	});
-
-	it('error changes to ENOENT mid-retry → re-queues (re-downloads), never mis-reports DISK_FULL', async () => {
-		const h = harness(0);
-		h.ds.writeChunkOutcomes = [
-			Object.assign(new Error('ENOSPC'), { code: 'ENOSPC' }), // initial write: disk full
-			Object.assign(new Error('ENOENT'), { code: 'ENOENT' }), // retry: file vanished during pause
-			null, // re-queued chunk re-downloaded, write now lands
-		];
-		await h.cd.run();
-
-		expect(h.client.requestChunkCalls).toBe(2); // re-queued → one extra fetch (blind retry would stay at 1)
-		expect(h.ds.downloadedChunks.has(h.chunkID)).toBe(true);
-		expect(h.errors).toHaveLength(0); // never burned through 5 retries reporting DISK_FULL
-	});
-
-	it('error changes to an unknown code mid-retry → fails with real cause, not DISK_FULL', async () => {
-		const h = harness(0);
-		h.ds.writeChunkOutcomes = [
-			Object.assign(new Error('ENOSPC'), { code: 'ENOSPC' }), // initial write: disk full
-			new Error('disk controller exploded'), // retry: unexpected error, no code
-		];
-		await h.cd.run();
-
-		expect(h.client.requestChunkCalls).toBe(1); // aborted immediately, no re-download
-		expect(h.errors).toHaveLength(1);
-		expect(h.errors[0]!.code).toBe(ErrorCodes.DOWNLOAD_ERROR); // real cause surfaced, not DISK_FULL
-		expect(h.ds.downloadedChunks.has(h.chunkID)).toBe(false);
-	});
-
-	it('two concurrent peers both hit ENOSPC → both write from memory, no extra fetches', async () => {
-		const mk = (seed: number) => {
-			const d = new Uint8Array(1024);
-			for (let i = 0; i < d.length; i++) d[i] = (i * seed + 7) & 0xff;
-			return d;
-		};
-		const dataA = mk(3);
-		const dataB = mk(11);
-		const idA = sha256Hex(dataA) as ChunkID;
-		const idB = sha256Hex(dataB) as ChunkID;
-
-		const ds = new MockDataServer();
-		// chunkSize must match the 1024-byte payloads — the downloader rejects chunk data
-		// whose length doesn't match the manifest, which would fail the write retry under test.
-		const lish = makeLISH({ chunkSize: 1024, files: [{ path: 'f.bin', size: dataA.length + dataB.length, checksums: [idA, idB] }] });
-		ds.add(lish);
-		ds.allChunkCount = 2;
-		ds.missingChunks = [makeMissingChunk(idA, 0, 0), makeMissingChunk(idB, 0, 1)];
-		// Disk full for the first 3 write attempts (2 initial + 1 owner retry), then it frees.
-		ds.writeChunkError = Object.assign(new Error('ENOSPC'), { code: 'ENOSPC' });
-		ds.writeChunkErrorCount = 3;
-
-		const chunkData = new Map<ChunkID, Uint8Array>([
-			[idA, dataA],
-			[idB, dataB],
-		]);
-		const c1 = new MockLISHClient();
-		c1.chunkData = chunkData;
-		const c2 = new MockLISHClient();
-		c2.chunkData = chunkData;
-
-		const peerManager = new PeerManager();
-		peerManager.setLishID(lish.id);
-		peerManager.tryAdd('peer-1', c1 as never, 'DIRECT');
-		peerManager.tryAdd('peer-2', c2 as never, 'DIRECT');
-
-		const state = { disabled: false, destroyed: false };
-		const pauseController = new PauseController(
-			() => state.disabled,
-			() => state.destroyed
-		);
-		const errors: Array<{ code: string; detail: string | undefined }> = [];
-		const deps: ChunkDownloaderDeps = {
-			lishID: lish.id,
-			downloadDir: '/tmp/test-dl',
-			abortSignal: new AbortController().signal,
-			dataServer: ds as never,
-			peerManager,
-			pauseController,
-			progressReporter: new ProgressReporter(),
-			fileAllocator: {} as never,
-			getLish: () => lish,
-			isDestroyed: () => state.destroyed,
-			isDisabled: () => state.disabled,
-			onSetError: (code, detail) => {
-				errors.push({ code, detail });
-				state.disabled = true;
-			},
-			onRetry: () => {},
-			emitAllocProgress: () => {},
-		};
-
-		await new ChunkDownloader(deps).run();
-
-		// Both chunks fetched exactly once total — the waiting peer wrote from its retained buffer.
-		expect(c1.requestChunkCalls + c2.requestChunkCalls).toBe(2);
-		expect(ds.downloadedChunks.has(idA)).toBe(true);
-		expect(ds.downloadedChunks.has(idB)).toBe(true);
-		expect(ds.writtenChunks).toHaveLength(2);
-		expect(errors).toHaveLength(0);
 	});
 });
